Safari-609.1.11.1/LayoutTests/canvas/philip/tests/2d.text.font.parse.size.percentage.default.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.font.parse.size.percentage.default</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var canvas2 = document.createElement('canvas');
var ctx2 = canvas2.getContext('2d');
ctx2.font = '1000% serif';
_assertSame(ctx2.font, '100px serif', "ctx2.font", "'100px serif'");


});
</script>
