LayoutTests/platform/mac-snowleopard/canvas/philip/tests/2d.text.font.parse.size.percentage.default-expected.txt

< [index] >
2d.text.font.parse.size.percentage.default
References: 2d.text.font.undefined
Actual output:
Failed assertion ctx.font === '100px serif' (got 1000% serif[string], expected 100px serif[string])
